Reference

​

Contents

​

Reference

Contents

Index

​

Index

​

DCISolver.DCIWorkspace

DCISolver.MetaDCI

DCISolver.SymCOOSolver

DCISolver.comp_λ_cgls

Base.success

DCISolver.TR_dogleg

DCISolver._compute_gradient_step!

DCISolver._compute_newton_step!

DCISolver._compute_step_length

DCISolver.compute_descent_direction!

DCISolver.compute_gBg

DCISolver.compute_lx!

DCISolver.dci

DCISolver.factorize!

DCISolver.feasibility_step

DCISolver.num_neg_eig

DCISolver.regularized_coo_saddle_system!

DCISolver.solve!

DCISolver.tangent_step!

​

DCISolver.DCIWorkspace — Type

DCIWorkspace(nlp, meta, x)

Pre-allocate the memory used during the dci call.

DCISolver.MetaDCI — Type

MetaDCI(x, y; kwargs...)

Structure containing all the parameters used in the dci call. x is an intial guess, and y is an initial guess for the Lagrange multiplier.

Arguments

The keyword arguments may include:

atol::T=T(1e-5): absolute tolerance.

rtol::T=T(1e-5): relative tolerance.

ctol::T=T(1e-5): feasibility tolerance.

unbounded_threshold::T=T(-1e5): below this threshold the problem is unbounded.

max_eval::Integer=50000: maximum number of cons + obj evaluations.

max_time::Float64=120.0: maximum number of seconds.

max_iter::Integer=500: maximum number of iterations.

max_iter_normal_step::Integer=typemax(Int): maximum number of iterations in normal step.

λ_struct::comp_λ_cgls=comp_λ_cgls(length(x0), length(y0), S).

linear_solver::Symbol=:ldlfact: Solver for the factorization. options: :ma57.

decrease_γ::T=T(0.1): Regularization for the factorization: reduce γ if possible, > √eps(T), between tangent steps.

increase_γ::T=T(100.0): Regularization for the factorization: up γ if possible, < 1/√eps(T), during the factorization.

δmin::T=√eps(T): Regularization for the factorization: smallest value of δ used for the regularization.

feas_step::Symbol=:feasibility_step: Normal step.

feas_η₁::T=T(1e-3): Feasibility step: decrease the trust-region radius when Ared/Pred < η₁.

feas_η₂::T=T(0.66): Feasibility step: increase the trust-region radius when Ared/Pred > η₂.

feas_σ₁::T=T(0.25): Feasibility step: decrease coefficient of the trust-region radius.

feas_σ₂::T=T(2.0): Feasibility step: increase coefficient of the trust-region radius.

feas_Δ₀::T=one(T): Feasibility step: initial radius.

bad_steps_lim::Integer=3: Feasibility step: consecutive bad steps before using a second order step.

feas_expected_decrease::T=T(0.95): Feasibility step: bad steps are when ‖c(z)‖ / ‖c(x)‖ >feas_expected_decrease.

TR_compute_step::Symbol=:TR_lsmr: Compute the direction in feasibility step: options: :TR_dogleg.

TR_struct::Union{TR_lsmr_struct, TR_dogleg_struct}=TR_lsmr_struct(length(x0), length(y0), S).

compρ_p1::T=T(0.75): update ρ as ρ = max(min(ngp, p1) * ρmax, ϵ).

compρ_p2::T=T(0.90): update ρ as ρ = primalnorm * p2 if not sufficiently feasible.

ρbar::T=T(2.0): radius of the larger cylinder is ρbar * ρ.

tan_Δ::T=one(T): Tangent step trust-region parameters: initial trust-region radius.

tan_η₁::T=T(1e-2): Tangent step trust-region parameters: decrease the trust-region radius when Ared/Pred < η₁.

tan_η₂::T=T(0.75): Tangent step trust-region parameters: increase the trust-region radius when Ared/Pred > η₂.

tan_σ₁::T=T(0.25): Tangent step trust-region parameters: decrease coefficient of the trust-region radius.

tan_σ₂::T=T(2.0): Tangent step trust-region parameters: increase coefficient of the trust-region radius.

tan_small_d::T=eps(T): Tangent step trust-region parameters: ||d|| is too small.

increase_Δtg::T=10: Tangent step trust-region parameters: increase if possible, < 1 / √eps(T), the Δtg between tangent steps.

For more details, we refer to the package documentation fine-tuneDCI.md.

DCISolver.SymCOOSolver — Type

An SymCOOSolver is an interface to allow simple usage of different solvers. Ideally, having rows, cols, vals and the dimension ndim of a symmetric matrix should allow the user to call M = LinearSolver(ndim, rows, cols, vals) factorize!(M) solve!(x, M, b) # Also x = M \ b Only the lower triangle of the matrix should be passed.

DCISolver.comp_λ_cgls — Type

comp_λ_cgls{<: AbstractFloat} attributes correspond to input parameters of cgls used in the computation of Lagrange multipliers.

Base.success — Function

success(M :: SymCOOSolver)

Returns whether factorize!(M) was successful.

DCISolver.TR_dogleg — Method

feasibility_step(nls, x, cx, Jx)

Compute a direction d such that min ‖cₖ + Jₖd‖ s.t. ||d|| ≤ Δ using a dogleg.

Also checks if problem is infeasible.

Returns 4 entries: (d, Jd, solved, infeasible)

DCISolver._compute_gradient_step! — Method

Compute a solution to minα q(-α g) s.t. ‖αg‖2 ≤ Δ

return dcp_on_boundary true if ‖αg‖ = Δ, return dcp = - α g return dcpBdcp = α^2 gBg and α the solution.

DCISolver._compute_newton_step! — Method

1st idea: we stop whenever the fonction looks convex-ish

DCISolver._compute_step_length — Method

Given two directions dcp and dn, compute the largest 0 ≤ τ ≤ 1 such that ‖dn + τ (dcp -dn)‖ = Δ

DCISolver.compute_descent_direction! — Method

Compute a direction d with three possible outcomes:

:cauchy_step

:newton

:dogleg

:interior_cauchy_step when γ is too large.

for min_d q(d) s.t. ‖d‖ ≤ Δ.

DCISolver.compute_gBg — Method

compute_gBg B is a symmetric sparse matrix whose lower triangular given in COO: (rows, cols, vals)

Compute ∇ℓxλ' * B * ∇ℓxλ

DCISolver.compute_lx! — Method

Compute the solution of ‖Jx' λ - ∇fx‖

DCISolver.dci — Method

dci(nlp; kwargs...)
dci(nlp, x; kwargs...)
dci(nlp, meta, x)

Compute a local minimum of an equality-constrained optimization problem nlp<:AbstractNLPModel, see NLPModels.jl. The returned value is a GenericExecutionStats, see SolverCore.jl.

If x is not specified, then nlp.meta.x0 is used as an initial guess. The keyword arguments are used to initialize a MetaDCI.

For advanced usage, the principal call to the solver uses a DCIWorkspace.

dci(nlp, meta, workspace)

This method implements the Dynamic Control of Infeasibility for equality-constrained problems described in

Dynamic Control of Infeasibility in Equality Constrained Optimization
Roberto H. Bielschowsky and Francisco A. M. Gomes
SIAM J. Optim., 19(3), 1299–1325.
https://doi.org/10.1137/070679557

Examples

julia> using DCISolver, ADNLPModels
julia> nlp = ADNLPModel(x -> 100 * (x[2] - x[1]^2)^2 + (x[1] - 1)^2, [-1.2; 1.0]);
julia> stats = dci(nlp)

DCISolver.factorize! — Function

factorize!(M :: SymCOOSolver)

Calls the factorization of the symmetric solver given by M. Use success(M) to check whether the factorization was successful.

DCISolver.feasibility_step — Method

feasibility_step(nls, x, cx, Jx)

Approximately solves min ‖c(x)‖.

Given xₖ, finds min ‖cₖ + Jₖd‖

DCISolver.num_neg_eig — Function

num_neg_eig(M :: SymCOOSolver)

Returns the number of negative eigenvalues of M.

DCISolver.regularized_coo_saddle_system! — Method

regularized_coo_saddle_system!(nlp, rows, cols, vals, γ = γ, δ = δ) Compute the structure for the saddle system [H + γI [Jᵀ]; J -δI] in COO-format in the following order: H J γ -δ

DCISolver.solve! — Function

solve!(x, M :: SymCOOSolver, b)

Solve the system $M x = b$. factorize!(M) should be called first.

DCISolver.tangent_step! — Method

min q(d):=¹/₂dᵀBd + dᵀg s.t Ad = 0 ‖d‖ ≦ Δ where B is an approximation of hessian of the Lagrangian, A is the jacobian matrix and g is the projected gradient.

Return status with outcomes:

:cauchy_step, :newton, :dogleg,

:unknown if we didn't enter the loop.

:smallhorizontalstep

:tired if we stop due to maxeval or maxtime

:success if we computed z such that ‖c(z)‖ ≤ meta.ρbar * ρ and Δℓ ≥ η₁ q(d)

See https://github.com/JuliaSmoothOptimizers/SolverTools.jl/blob/78f6793f161c3aac2234aee8a27aa07f1df3e8ee/src/trust-region/trust-region.jl#L37 for SolverTools.aredpred