Reference

​

Contents

​

Reference

Contents

Index

​

Index

​

DCISolver.SymCOOSolver

DCISolver.comp_λ_cgls

Base.success

DCISolver.TR_dogleg

DCISolver._compute_gradient_step

DCISolver._compute_newton_step!

DCISolver._compute_step_length

DCISolver.compute_descent_direction

DCISolver.compute_gBg

DCISolver.compute_lx

DCISolver.dci

DCISolver.factorize!

DCISolver.feasibility_step

DCISolver.num_neg_eig

DCISolver.regularized_coo_saddle_system!

DCISolver.solve!

DCISolver.tangent_step

​

DCISolver.SymCOOSolver — Type

An SymCOOSolver is an interface to allow simple usage of different solvers. Ideally, having rows, cols, vals and the dimension ndim of a symmetric matrix should allow the user to call M = LinearSolver(ndim, rows, cols, vals) factorize!(M) solve!(x, M, b) # Also x = M \ b Only the lower triangle of the matrix should be passed.

DCISolver.comp_λ_cgls — Type

comp_λ_cgls{<: AbstractFloat} attributes correspond to input parameters of cgls used in the computation of Lagrange multipliers.

Base.success — Function

success(M :: SymCOOSolver)

Returns whether factorize!(M) was successful.

DCISolver.TR_dogleg — Method

feasibility_step(nls, x, cx, Jx)

Compute a direction d such that min ‖cₖ + Jₖd‖ s.t. ||d|| ≤ Δ using a dogleg.

Also checks if problem is infeasible.

Returns 4 entries: (d, Jd, solved, infeasible)

DCISolver._compute_gradient_step — Method

Compute a solution to minα q(-α g) s.t. ‖αg‖2 ≤ Δ

return dcp_on_boundary true if ‖αg‖ = Δ, return dcp = - α g return dcpBdcp = α^2 gBg and α the solution.

DCISolver._compute_newton_step! — Method

1st idea: we stop whenever the fonction looks convex-ish

DCISolver._compute_step_length — Method

Given two directions dcp and dn, compute the largest 0 ≤ τ ≤ 1 such that ‖dn + τ (dcp -dn)‖ = Δ

DCISolver.compute_descent_direction — Method

Compute a direction d with three possible outcomes:

:cauchy_step

:newton

:dogleg

:interior_cauchy_step when γ is too large.

for min_d q(d) s.t. ‖d‖ ≤ Δ.

DCISolver.compute_gBg — Method

compute_gBg B is a symmetric sparse matrix whose lower triangular given in COO: (rows, cols, vals)

Compute ∇ℓxλ' * B * ∇ℓxλ

DCISolver.compute_lx — Method

Compute the solution of ‖Jx' λ - ∇fx‖

DCISolver.dci — Method

dci(nlp, x; kwargs...)

This method implements the Dynamic Control of Infeasibility for equality-constrained problems described in

Dynamic Control of Infeasibility in Equality Constrained Optimization
Roberto H. Bielschowsky and Francisco A. M. Gomes
SIAM J. Optim., 19(3), 1299–1325.
https://doi.org/10.1137/070679557

DCISolver.factorize! — Function

factorize!(M :: SymCOOSolver)

Calls the factorization of the symmetric solver given by M. Use success(M) to check whether the factorization was successful.

DCISolver.feasibility_step — Method

feasibility_step(nls, x, cx, Jx)

Approximately solves min ‖c(x)‖.

Given xₖ, finds min ‖cₖ + Jₖd‖

DCISolver.num_neg_eig — Function

num_neg_eig(M :: SymCOOSolver)

Returns the number of negative eigenvalues of M.

DCISolver.regularized_coo_saddle_system! — Method

regularized_coo_saddle_system!(nlp, rows, cols, vals, γ = γ, δ = δ) Compute the structure for the saddle system [H + γI [Jᵀ]; J -δI] in COO-format in the following order: H J γ -δ

DCISolver.solve! — Function

solve!(x, M :: SymCOOSolver, b)

Solve the system $M x = b$. factorize!(M) should be called first.

DCISolver.tangent_step — Method

min q(d):=¹/₂dᵀBd + dᵀg s.t Ad = 0 ‖d‖ ≦ Δ where B is an approximation of hessian of the Lagrangian, A is the jacobian matrix and g is the projected gradient.

Return status with outcomes:

:cauchy_step, :newton, :dogleg,

:unknown if we didn't enter the loop.

:smallhorizontalstep

:tired if we stop due to maxeval or maxtime

:success if we computed z such that ‖c(z)‖ ≤ 2ρ and Δℓ ≥ η₁ q(d)

See https://github.com/JuliaSmoothOptimizers/SolverTools.jl/blob/78f6793f161c3aac2234aee8a27aa07f1df3e8ee/src/trust-region/trust-region.jl#L37 for SolverTools.aredpred